requirements.txt
flask
google-cloud-aiplatform
google-cloud-storage
python-dotenv
pillow
opencv-python-headless